+++ CHANGELOG.md
@@ -26,7 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `setColorOnLedStrip` GraphQL mutation changes `color` for a certain ledStrip.
 - `setAnimationTypeOnLedStrip` GraphQL mutation changes `animationType` for a certain ledStrip.
 - `setAnimationTypeOnLedStrip` GraphQL mutation changes `animationType` for a certain ledStrip.
-- `devices` GraphQL query - lets a user query `Device`s on an array of `secret`s.
+- `devices` GraphQL query lets a user query `Device`s on an array of `secret`s.
+- `secretIsValid` GraphQL query lets a user check if a secret belongs to a `Device`.
 
 ### Changed